Prefs clock: split up/reorganize ClockView
[code written in 2004, but was never committed because CVS sucks]

The ClockView class just never sat right with me; it has two different
display modes, and the two really have nothing to do with each other
except that they're both recognizeable as "clocks". At the same time,
Banlu Kemiyatorn (aka "]d") was working on a new thing that eventually
became known as "AClock", and I wanted to support it.

Well, here's what I wrote for that kind of thing. It all builds (or at
least built last time I recompiled GNUstep/Backbone), but it doesn't work
because there's no way to set a clock delegate in the code.

The reason I'm checking this in is that I've got a new idea for clock
handling, and of course I want the code in revision control before that
happens.

Here's the plan:
    1. Create the Backbone system lib/fw that we always talked about.
    2. Add BBClockView (and other classes, obviously) to it, clean it
       up, add defaults etc. so there's a functional default that looks
       basically similar to the NeXT calendar-clock.
    3. Define a bundle fmt for describing custom clock views, and redefine
       the current classes to use them.
    4. ???
    5. Profit!