TextEdit: handle NSNotificationCenter exceptions
So, we hit the perfect storm of goofiness.

Apparently GNUstep's NSScrollView class doesn't implement the
-fixUpScrollViewBackgroundColor: method (first implemented in OPENSTEP 4.1,
and used to make a scroll view notice when the system colors have changed),
and NSNotificationCenter has been changed to throw exceptions rather than
silently deal. As a result, when TextEdit tries to ask to be informed when
the colors have changed (so NSScrollView, via ScalingScrollView, can update
the colors in the scrollbar), the notification center throws an exception
and the runloop goes apeshit.

Blah.

Fix the bug by just ignoring the exception as it should have been anyway.