Start using autoconf for dealing with portability issues.
It's pretty common knowledge that I never applied Riccardo Mottola's
Solaris-compatibility patch because of how it was written...that is,
it was essentially specific to Solaris and would require anyone on another
System V-derived machine to modify the code, and that is deeply ugly. He
disagreed, and we had a bit of an impasse because "my way" wasn't a
particularly easy fix.

Until now. :)