Preferences: Nuke the ancillary makefiles
Move the custom makefile code out of the .preamble & .postamble files, and
delete them. Anyone who knows how to use them ought to be clever enough
to create them in the first place, so we're either confusing people who
don't know what they're for, or getting in the way of people who do.

Either way, we shouldn't include them in the repo.