Disable hyphenation on GNUstep
I'm sick of seeing the hyphenation warnings every time I start TextEdit,
in particular because TextEdit actually has code to detect if hyphenation
(which was added to OPENSTEP in version 4.2) is supported. According to the
code, GNUstep supports hyphenation...but then it gives a warning every time
you try to use it. Grr.