open: Improve stdin -> app support
Use mkstemps() to make our temp file instead of just generating a name.
Also, put a category on NSData to write to a file descriptor so we can
write stdin to the file descriptor mkstemps() hands us.

The reason we need mkstemps instead of just plain-old mkstemp is that our
temp files currently need to have extensions. One day, we should be
able to generate NSFileWrapper objects so we don't have to use temp files,
but this is not that day.

I'll try to create or import a version that will work on systems that don't
have it.