open: start supporting .desktop files
Eventually, the BBAppKit framework will include an INI-style file reader
(and perhaps writer) and Workspace will use it to allow integration with
non-GNUstep-based apps... but for now, this is a REALLY dirty (seriously,
don't look at it, you may go blind) .desktop file reader that is just
barely good enough to handle doing "open /etc/xdg/autostart/*.desktop"
when logging in.

It can't deal properly with a number of the .desktop files you'll find in a
typical /usr/share/applications dir (because it just robotically reads all
of the keys it understands), so .desktop files with multiple stanzas
confuse the hell out of it.

Like I said, it's temporary and sucks, but it's still reasonably useful. :)