make BBVCTextField work in Gorm
Even if you subclass NSTextField in Gorm and set a text field to be of
class 'BBVCTextField', Gorm is still encoding text fields with the old
cell class. Make BBVCTextField force the issue in -awakeFromNib by cloning
the encoded cell onto a BBVCTextFieldCell.

I need to look into building a palette so we don't have to subclass in Gorm
at all.