Terminal: Add support for bright background color codes, xterm (kinda)
The latest Linux kernels have changed some stuff in the terminal driver,
especially with regard to the SGR (CSI m) command. They have added support
for setting the foreground and background to colors 8-15 directly, using
the 90-97 and 100-107 codes; they have changed the meaning of code 21 to
enable underline instead of set to normal intensity, and changed the
meaning of code 38 from "bright white with underline" to "set xterm-style
foreground".

As it turns out, not understanding these codes is "dangerous" even if you
don't support them. Terminal did not understand the 48 code and understood
38 differently, and as it should, it ignored them. But these are _compound_
codes, and are followed by more that depend on them -- some of which are
interpreted differently than expected.

The xterm codes CSI [34]8 ; 5 ; n m selects xterm 256 color n -- if n is
less than 16, it's pretty much the same as selecting one of our palette
entries, only they're specific colors not subject to palette altering. The
rest of the colors are a 3x3x3 color cube and a 16-level greyscale, but we
can't support those at this time so we ignore them.

The code CSI [34]8 ; 2 ; r ; g ; b m is a 24-bit version, for selecting
specific colors, where r g and b are each a number 0-255. We can't support
this either, so we understand and pointedly ignore it.